+++ server/app/Http/Controllers/Posts/VideoController.php
@@ -7,24 +7,28 @@ use App\Http\Requests\Posts\StoreVideoRequest;
 use App\Http\Requests\Posts\UpdateVideoRequest;
 use App\Models\Posts\Video;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class VideoController extends Controller
 {
     public function index()
 {
+    Gate::authorize('viewAny',Video::class);
     $videos = Video::all();
     return response()->json($videos);
 }
 
 public function show(Video $video)
 {
+    Gate::authorize('view',Video::class);
     return response()->json($video);
 }
 
     
     public function store(StoreVideoRequest $request)
     {
+        Gate::authorize('create',Video::class);
         $video = Video::create($request->validated());
 
         return response()->json([
@@ -35,6 +39,7 @@ public function show(Video $video)
 
     public function update(UpdateVideoRequest $request, Video $video)
     {
+        Gate::authorize('update',Video::class);
         $video->update($request->validated());
         
 
@@ -46,6 +51,7 @@ public function show(Video $video)
 
     public function destroy(Video $video)
     {
+        Gate::authorize('elete',Video::class);
 
         $video->delete();
